Implementation
A exact top-k retreival system is constructed and the program accepts the collection, processes the documents and queries.
The program accepts a free text query and calculates the cosine similarity score for the documents. The program displays the names of top k documents for every query in decreasing order. Same is implemented using the inexact retreival and results are compared.